refactor(k8s): reworked deployment to use k3s instead of k0s
k0s proved to be more difficult in the long run when working with raspberry pis. moving to the
standard k3s deployment using ansible seemed like the right move to make considering that k3s is
tested and proven on raspberry pi and the setup is just as easy. `gilt` is being used to pull in a
playbook so that i dont have to clutter this repo any more than it already is. i've also added more
directions on how to set the cluster up on the initial go. good luck and dog bless.

BREAKING CHANGE: k0s no longer works, long live k3s

#61